+++ src/strife/r_plane.h
@@ -26,7 +26,7 @@
 
 
 // Visplane related.
-extern  short*		lastopening;
+extern  int*		lastopening; // [crispy] 32-bit integer math
 
 
 typedef void (*planefunction_t) (int top, int bottom);
@@ -34,8 +34,8 @@ typedef void (*planefunction_t) (int top, int bottom);
 extern planefunction_t	floorfunc;
 extern planefunction_t	ceilingfunc_t;
 
-extern short		floorclip[MAXWIDTH];
-extern short		ceilingclip[MAXWIDTH];
+extern int		floorclip[MAXWIDTH]; // [crispy] 32-bit integer math
+extern int		ceilingclip[MAXWIDTH]; // [crispy] 32-bit integer math
 
 extern fixed_t		yslope[MAXHEIGHT];
 extern fixed_t		distscale[MAXWIDTH];
@@ -52,10 +52,10 @@ R_MapPlane
 void
 R_MakeSpans
 ( int		x,
-  int		t1,
-  int		b1,
-  int		t2,
-  int		b2 );
+  unsigned int		t1, // [crispy] 32-bit integer math
+  unsigned int		b1, // [crispy] 32-bit integer math
+  unsigned int		t2, // [crispy] 32-bit integer math
+  unsigned int		b2 ); // [crispy] 32-bit integer math
 
 void R_DrawPlanes (void);
