+++ src/strife/r_things.c
@@ -75,8 +75,8 @@ lighttable_t**	spritelights;
 
 // constant arrays
 //  used for psprite clipping and initializing clipping
-short		negonearray[MAXWIDTH];
-short		screenheightarray[MAXWIDTH];
+int		negonearray[MAXWIDTH]; // [crispy] 32-bit integer math
+int		screenheightarray[MAXWIDTH]; // [crispy] 32-bit integer math
 
 
 //
@@ -338,8 +338,8 @@ vissprite_t* R_NewVisSprite (void)
 // Masked means: partly transparent, i.e. stored
 //  in posts/runs of opaque pixels.
 //
-short*		mfloorclip;
-short*		mceilingclip;
+int*		mfloorclip; // [crispy] 32-bit integer math
+int*		mceilingclip; // [crispy] 32-bit integer math
 
 fixed_t		spryscale;
 int64_t		sprtopscreen; // [crispy] WiggleFix
@@ -980,8 +980,8 @@ void R_SortVisSprites (void)
 void R_DrawSprite (vissprite_t* spr)
 {
     drawseg_t*		ds;
-    short		clipbot[MAXWIDTH];
-    short		cliptop[MAXWIDTH];
+    int		clipbot[MAXWIDTH]; // [crispy] 32-bit integer math
+    int		cliptop[MAXWIDTH]; // [crispy] 32-bit integer math
     int			x;
     int			r1;
     int			r2;
